Add collision test debugging screen. Fixed issue with initial choice screen. Added cubic bezier function to math utilities.

+++ src/Game/DefaultECPS/generalProcesses.c
@@ -730,7 +730,7 @@ bool collisionTestAndResponse( Entity* eOne, Entity* eTwo, ComponentID firstComp
 	return false;
 }
 
-void collisionResponse( int firstColliderIdx, int secondColliderIdx, Vector2 separation )
+void collisionResponse( size_t firstColliderIdx, size_t secondColliderIdx, Vector2 separation )
 {
 	Entity eOne;
 	Entity eTwo;
